+++ CHANGES.rst
@@ -5,7 +5,8 @@ History of xm.tracker
 1.0.7 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Never cache the tracker view.  It always needs to be fresh.
+  [maurits]
 
 
 1.0.6 (2012-09-12)